+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Fixed potential race during `Conn.Close()`.
 * Disable sending frames when closing `Session`, `Sender`, and `Receiver`.
 * Don't leak in-flight messages when a message settlement API is cancelled or times out waiting for acknowledgement.
+* `Sender.Send()` will return an `*amqp.Error` with condition `amqp.ErrCondTransferLimitExceeded` when attempting to send a transfer on a link with no credit.
 
 ### Other Changes